pm-page: unwind circular unit ref.
